Add support for functions containing lambda expressions
Now a function definition needs to start with an access modifier
and/or a type, and have a `(` in the first line of the signature,
before the scope opening (`{` or `=>`). This prevents the text object
from matching most lambda expressions, but also namespace/class
definitions which it used to (before this commit).