Update text object naming convention
Switch the behavior of a/i and OBJECT/object:

Before:

a:  with blank        OBJECT: outside
   lines/spaces

i: without blank      object: inside
   lines/spaces

Now:

a:   outside          OBJECT:  with blank
                              lines/spaces

i:   inside           object: without blank
     (inner)                   lines/spaces

This depends on the names you choose to the text objects,
it is impossible to come up with a convention that satisfies all.

Now it seems it's incomplete, there's no "i" text objects.
But at least we won't be building muscle memory for something
temporary.